Changed the node_id type to uint32 to save space

+++ bplustree_utils.c
@@ -4,11 +4,11 @@ internal void bpt_print_leaf( node* n, int indent ) {
 	char ind[100] = "                               END";
 	ind[indent*2] = '\0';
 	
-	printf("%sL(%lu)-[ ", ind, n->id);
+	printf("%sL(%u)-[ ", ind, n->id);
 	for( size_t i=0; i<n->num_keys; i++ ) {
 		printf("0x%lx ", n->keys[i] );
 	}
-	printf("] - (%p) (#keys: %lu, parent id %lu)\n", (void *)n, n->num_keys, n->parent_node_id);
+	printf("] - (%p) (#keys: %u, parent id %u)\n", (void *)n, n->num_keys, n->parent_node_id);
 
 }
 
@@ -23,11 +23,11 @@ internal inline void bpt_print( database* db, node* start, int indent ) {
 		return;
 	}
 
-	printf("%sN(%lu) (%p) keys: %zu (parent id %lu)\n", ind, start->id, (void *)start, start->num_keys, start->parent_node_id);
+	printf("%sN(%u) (%p) keys: %u (parent id %u)\n", ind, start->id, (void *)start, start->num_keys, start->parent_node_id);
 		
 	// print every key/node
 	node* n;
-	for( size_t i=0; i<start->num_keys; i++ ) {
+	for( uint32 i=0; i<start->num_keys; i++ ) {
 		n = retrieve_node( db, start->pointers[i].child_node_id  );
 		assert( n != NULL );
 
@@ -36,7 +36,7 @@ internal inline void bpt_print( database* db, node* start, int indent ) {
 		} else {
 			bpt_print( db, n, indent+1 );			
 		}
-		printf("%sK[%lu]-[ 0x%lx ]\n", ind, i, start->keys[i] );
+		printf("%sK[%u]-[ 0x%lx ]\n", ind, i, start->keys[i] );
 		release_node( db, n );
 		
 	}
@@ -64,25 +64,25 @@ internal inline board63 max_key( database* db, node* n ) {
 #ifdef VERBOSE	
 internal void check_tree_correctness( database* db, node* n ) {
 
-	print("Checking correctness of node %lu", n->id );
+	print("Checking correctness of node %u", n->id );
 	
 	// for leaves, check if all keys are ascending
 	if( n->is_leaf ) {
 		
-		for(size_t i=0; i<n->num_keys-1; i++ ) {
+		for(uint32 i=0; i<n->num_keys-1; i++ ) {
 			assert( n->keys[i] < n->keys[i+1]);
 		}
-		print("node %lu OK", n->id);
+		print("node %u OK", n->id);
 		return;
 	}
 	
 
-	for(size_t i=0; i<n->num_keys; i++ ) {
+	for(uint32 i=0; i<n->num_keys; i++ ) {
 		// for every key, the max
 		node* temp = retrieve_node( db, n->pointers[i].child_node_id ); 
 		board63 max = max_key( db, temp );
 		release_node( db, temp );
-		print("key[%lu] = 0x%lx, max from left node = 0x%lx", i, n->keys[i], max );
+		print("key[%u] = 0x%lx, max from left node = 0x%lx", i, n->keys[i], max );
 		assert( max < n->keys[i] );
 	}
 	
@@ -90,9 +90,9 @@ internal void check_tree_correctness( database* db, node* n ) {
 	node* temp = retrieve_node( db, n->pointers[ n->num_keys ].child_node_id ); 
 	board63 max = max_key( db, temp );
 	release_node( db, temp );
-	print("key[%lu] = 0x%lx, max from right node = 0x%lx", n->num_keys-1, n->keys[ n->num_keys-1], max );
+	print("key[%u] = 0x%lx, max from right node = 0x%lx", n->num_keys-1, n->keys[ n->num_keys-1], max );
 	assert( n->keys[ n->num_keys-1] <= max );
-	print("node %lu OK", n->id);
+	print("node %u OK", n->id);
 }
 #else
 #define check_tree_correctness( db, n ) // nothing
